+++ Sources/PresentationKit/Navigation/NavigationChevron.swift
@@ -13,32 +13,54 @@ public struct NavigationChevron: View {
 
     public init() {}
 
+        @Environment(\.navigationChevronStyle) var style
+
     public var body: some View {
         Image(systemName: "chevron.right")
-            .font(font)
-            .opacity(opacity)
-            .padding(.leading, padding)
-            .scaleEffect(scale)
+            .font(style.font)
+            .foregroundStyle(style.color)
+            .padding(.leading, style.padding)
+            .scaleEffect(style.scale)
     }
 }
 
-extension NavigationChevron {
+// MARK: - Style
 
-    var font: Font {
-        #if os(iOS)
-        Font.footnote.weight(.semibold)
-        #elseif os(tvOS)
+public struct NavigationChevronStyle {
+    public init(
+        font: Font = standardFont,
+        color: Color = .secondary.opacity(0.6),
+        padding: Double = 2,
+        scale: Double = standardScale
+    ) {
+        self.font = font
+        self.color = color
+        self.padding = padding
+        self.scale = scale
+    }
+
+    public var font: Font
+    public var color: Color
+    public var padding: Double
+    public var scale: Double
+}
+
+public extension NavigationChevronStyle {
+
+    static var standard: Self { .init() }
+}
+
+public extension NavigationChevronStyle {
+
+    static var standardFont: Font {
+        #if os(tvOS)
         Font.caption.weight(.bold)
         #else
         Font.footnote.weight(.semibold)
         #endif
     }
-    
-    var opacity: Double { 0.25 }
 
-    var padding: Double { 2 }
-
-    var scale: Double {
+    static var standardScale: Double {
         #if os(iOS)
         1.05
         #elseif os(tvOS)
@@ -46,5 +68,19 @@ extension NavigationChevron {
         #else
         1.00
         #endif
+    }
+}
+
+// MARK: - Environment
+
+private extension EnvironmentValues {
+
+    @Entry var navigationChevronStyle: NavigationChevronStyle = .standard
+}
+
+public extension View {
+
+    func navigationChevronStyle(_ style: NavigationChevronStyle) -> some View {
+        environment(\.navigationChevronStyle, style)
     }
 }